+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Breaking changes
 * Logging can now be configured through driver options. However, the environment variables for logging are no longer evaluated lazily, so they now need to be set before requiring `zwave-js`.
 * The second (string) parameter of the `"interview failed"` event handler was removed
+* The type `ValueMetadataBase` has been renamed to `ValueMetadataAny`. The old type `ValueMetadataAny` was merged into `ValueMetadataBase`.
 
 ### Config file changes
 * Added a config file for `HeatIt Z-TRM3`